requirements.txt
discord.py
groq